# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the docs of Bikes Compass API from Bikes2Road.",
        "title": "Bikes Compass API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
        "/bikes/byke/{hash_byke}": {
            "get": {
                "description": "This service extract all data from a Byke by Hash_Byke",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bikes 2 Road"
                ],
                "summary": "Search Byke by Hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hash of Byke that you want extract",
                        "name": "hash_byke",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.GetBykeResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseHttpError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseHttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseHttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseHttpError"
                        }
                    }
                }
            }
        },
        "/bikes/placeholder": {
            "get": {
                "description": "This service extract a list of name bikes from a Byke by name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bikes 2 Road"
                ],
                "summary": "Search Byke by Hash",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Yamaha",
                        "description": "name of byke that you want search",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PlaceHolderResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseHttpError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseHttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseHttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseHttpError"
                        }
                    }
                }
            }
        },
        "/bikes/search": {
            "get": {
                "description": "This service extract all bikes with pagination, you can search bikes by name, or all bikes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bikes 2 Road"
                ],
                "summary": "Search Bikes",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "page that you want extract",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 30,
                        "type": "integer",
                        "description": "cant bikes you want extract",
                        "name": "cant",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "BMW M1000RR",
                        "description": "name of byke that you want search",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "BMW",
                        "description": "brand of byke that you want search",
                        "name": "brand",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.GetAllResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseHttpError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseHttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseHttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ResponseHttpError"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "This service returns OK status to verify the microservice is running",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.HealthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.GetAllResponseSuccess": {
            "type": "object",
            "required": [
                "data",
                "success",
                "total"
            ],
            "properties": {
                "data": {
                    "description": "Lista de motos encontradas",
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "success": {
                    "description": "Indica si la petición fue exitosa",
                    "type": "boolean",
                    "example": true
                },
                "total": {
                    "description": "Número total de registros encontrados",
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "domain.GetBykeResponseSuccess": {
            "type": "object",
            "required": [
                "data",
                "success",
                "total"
            ],
            "properties": {
                "data": {
                    "description": "Lista de motos encontradas",
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "success": {
                    "description": "Indica si la petición fue exitosa",
                    "type": "boolean",
                    "example": true
                },
                "total": {
                    "description": "Número total de registros encontrados",
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "domain.HealthResponse": {
            "type": "object",
            "required": [
                "message",
                "success"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "OK"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "domain.PlaceHolderResponseSuccess": {
            "type": "object",
            "required": [
                "data",
                "success",
                "total"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"Ducati\"",
                        "\"BMW\"]"
                    ]
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "total": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "domain.ResponseHttpError": {
            "type": "object",
            "required": [
                "code",
                "error",
                "message",
                "success"
            ],
            "properties": {
                "code": {
                    "description": "Código de error HTTP",
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "description": "Identificador del tipo de error",
                    "type": "string",
                    "example": "bad_request"
                },
                "message": {
                    "description": "Mensaje descriptivo del error",
                    "type": "string",
                    "example": "error with request"
                },
                "success": {
                    "description": "Indica si la operación fue exitosa (en este caso false)",
                    "type": "boolean",
                    "example": false
                }
            }
        }
    }
}